refactor(checkout): move ticket selector to separate component
