+++ src/components/TicketSelector/index.js
@@ -1,0 +1,1 @@
+export { TicketSelector } from './TicketSelector';
